// FabricExample/.yalc/react-native-plaid-link-sdk/src/Types.ts
export enum LinkAccountType {
    CREDIT = 'credit',
    DEPOSITORY = 'depository',
    INVESTMENT = 'investment',
    LOAN = 'loan',
    OTHER = 'other',
}
// ...
export interface LinkAccountSubtype {
}
// ...
export interface LinkSuccess {
    publicToken: string;
    metadata: LinkSuccessMetadata;
}

export interface LinkSuccessMetadata {
    institution?: LinkInstitution;
    accounts: LinkAccount[];
    linkSessionId: string;
    metadataJson?: string;
}

export interface LinkAccount {
    id: string;
    name?: string;
    mask?: string;
    type: LinkAccountType;
    subtype: LinkAccountSubtype;
    verificationStatus?: LinkAccountVerificationStatus;
}

export enum LinkAccountVerificationStatus {
    PENDING_AUTOMATIC_VERIFICATION = 'pending_automatic_verification',
    PENDING_MANUAL_VERIFICATION = 'pending_manual_verification',
    MANUALLY_VERIFIED = 'manually_verified',
}

export interface LinkInstitution {
    id: string;
    name: string;
}

export interface LinkExit {
    error?: LinkError;
    metadata: LinkExitMetadata;
}

export interface LinkExitMetadata {
    status?: LinkExitMetadataStatus;
    institution?: LinkInstitution;
    linkSessionId: string;
    requestId: string;
    metadataJson?: string;
}
// ...
export interface LinkError {
    errorCode: LinkErrorCode;
    errorType: LinkErrorType;
    errorMessage: string;
    /** @deprecated DO NOT USE, data not guaranteed. Use `displayMessage` instead */
    errorDisplayMessage?: string;
    displayMessage?: string;
    errorJson?: string;
}

export enum LinkErrorCode {
    // ITEM_ERROR
    INVALID_CREDENTIALS = "INVALID_CREDENTIALS",
    INVALID_MFA = "INVALID_MFA",
    ITEM_LOGIN_REQUIRED = "ITEM_LOGIN_REQUIRED",
    INSUFFICIENT_CREDENTIALS = "INSUFFICIENT_CREDENTIALS",
    ITEM_LOCKED = "ITEM_LOCKED",
    USER_SETUP_REQUIRED = "USER_SETUP_REQUIRED",
    MFA_NOT_SUPPORTED = "MFA_NOT_SUPPORTED",
    INVALID_SEND_METHOD = "INVALID_SEND_METHOD",
    NO_ACCOUNTS = "NO_ACCOUNTS",
    ITEM_NOT_SUPPORTED = "ITEM_NOT_SUPPORTED",
    TOO_MANY_VERIFICATION_ATTEMPTS = "TOO_MANY_VERIFICATION_ATTEMPTS",

    INVALD_UPDATED_USERNAME = "INVALD_UPDATED_USERNAME",
    INVALID_UPDATED_USERNAME = "INVALID_UPDATED_USERNAME",

    ITEM_NO_ERROR = "ITEM_NO_ERROR",
    item_no_error = "item-no-error",
    NO_AUTH_ACCOUNTS = "NO_AUTH_ACCOUNTS",
    NO_INVESTMENT_ACCOUNTS = "NO_INVESTMENT_ACCOUNTS",
    NO_INVESTMENT_AUTH_ACCOUNTS = "NO_INVESTMENT_AUTH_ACCOUNTS",
    NO_LIABILITY_ACCOUNTS = "NO_LIABILITY_ACCOUNTS",
    PRODUCTS_NOT_SUPPORTED = "PRODUCTS_NOT_SUPPORTED",
    ITEM_NOT_FOUND = "ITEM_NOT_FOUND",
    ITEM_PRODUCT_NOT_READY = "ITEM_PRODUCT_NOT_READY",

    // INSTITUTION_ERROR
    INSTITUTION_DOWN = "INSTITUTION_DOWN",
    INSTITUTION_NOT_RESPONDING = "INSTITUTION_NOT_RESPONDING",
    INSTITUTION_NOT_AVAILABLE = "INSTITUTION_NOT_AVAILABLE",
    INSTITUTION_NO_LONGER_SUPPORTED = "INSTITUTION_NO_LONGER_SUPPORTED",

    // API_ERROR
    INTERNAL_SERVER_ERROR = "INTERNAL_SERVER_ERROR",
    PLANNED_MAINTENANCE = "PLANNED_MAINTENANCE",

    // ASSET_REPORT_ERROR
    PRODUCT_NOT_ENABLED = "PRODUCT_NOT_ENABLED",
    DATA_UNAVAILABLE = "DATA_UNAVAILABLE",
    ASSET_PRODUCT_NOT_READY = "ASSET_PRODUCT_NOT_READY",
    ASSET_REPORT_GENERATION_FAILED = "ASSET_REPORT_GENERATION_FAILED",
    INVALID_PARENT = "INVALID_PARENT",
    INSIGHTS_NOT_ENABLED = "INSIGHTS_NOT_ENABLED",
    INSIGHTS_PREVIOUSLY_NOT_ENABLED = "INSIGHTS_PREVIOUSLY_NOT_ENABLED",

    // BANK_TRANSFER_ERROR
    BANK_TRANSFER_LIMIT_EXCEEDED = "BANK_TRANSFER_LIMIT_EXCEEDED",
    BANK_TRANSFER_MISSING_ORIGINATION_ACCOUNT = "BANK_TRANSFER_MISSING_ORIGINATION_ACCOUNT",
    BANK_TRANSFER_INVALID_ORIGINATION_ACCOUNT = "BANK_TRANSFER_INVALID_ORIGINATION_ACCOUNT",
    BANK_TRANSFER_ACCOUNT_BLOCKED = "BANK_TRANSFER_ACCOUNT_BLOCKED",
    BANK_TRANSFER_INSUFFICIENT_FUNDS = "BANK_TRANSFER_INSUFFICIENT_FUNDS",
    BANK_TRANSFER_NOT_CANCELLABLE = "BANK_TRANSFER_NOT_CANCELLABLE",
    BANK_TRANSFER_UNSUPPORTED_ACCOUNT_TYPE = "BANK_TRANSFER_UNSUPPORTED_ACCOUNT_TYPE",
    BANK_TRANSFER_UNSUPPORTED_ENVIRONMENT = "BANK_TRANSFER_UNSUPPORTED_ENVIRONMENT",

    // SANDBOX_ERROR
    SANDBOX_PRODUCT_NOT_ENABLED = "SANDBOX_PRODUCT_NOT_ENABLED",
    SANDBOX_WEBHOOK_INVALID = "SANDBOX_WEBHOOK_INVALID",
    SANDBOX_BANK_TRANSFER_EVENT_TRANSITION_INVALID = "SANDBOX_BANK_TRANSFER_EVENT_TRANSITION_INVALID",

    // INVALID_REQUEST
    MISSING_FIELDS = "MISSING_FIELDS",
    UNKNOWN_FIELDS = "UNKNOWN_FIELDS",
    INVALID_FIELD = "INVALID_FIELD",
    INCOMPATIBLE_API_VERSION = "INCOMPATIBLE_API_VERSION",
    INVALID_BODY = "INVALID_BODY",
    INVALID_HEADERS = "INVALID_HEADERS",
    NOT_FOUND = "NOT_FOUND",
    NO_LONGER_AVAILABLE = "NO_LONGER_AVAILABLE",
    SANDBOX_ONLY = "SANDBOX_ONLY",
    INVALID_ACCOUNT_NUMBER = "INVALID_ACCOUNT_NUMBER",

    // INVALID_INPUT
    // From above ITEM_LOGIN_REQUIRED = "INVALID_CREDENTIALS",
    INCORRECT_DEPOSIT_AMOUNTS = "INCORRECT_DEPOSIT_AMOUNTS",
    UNAUTHORIZED_ENVIRONMENT = "UNAUTHORIZED_ENVIRONMENT",
    INVALID_PRODUCT = "INVALID_PRODUCT",
    UNAUTHORIZED_ROUTE_ACCESS = "UNAUTHORIZED_ROUTE_ACCESS",
    DIRECT_INTEGRATION_NOT_ENABLED = "DIRECT_INTEGRATION_NOT_ENABLED",
    INVALID_API_KEYS = "INVALID_API_KEYS",
    INVALID_ACCESS_TOKEN = "INVALID_ACCESS_TOKEN",
    INVALID_PUBLIC_TOKEN = "INVALID_PUBLIC_TOKEN",
    INVALID_LINK_TOKEN = "INVALID_LINK_TOKEN",
    INVALID_PROCESSOR_TOKEN = "INVALID_PROCESSOR_TOKEN",
    INVALID_AUDIT_COPY_TOKEN = "INVALID_AUDIT_COPY_TOKEN",
    INVALID_ACCOUNT_ID = "INVALID_ACCOUNT_ID",
    MICRODEPOSITS_ALREADY_VERIFIED = "MICRODEPOSITS_ALREADY_VERIFIED",

    // INVALID_RESULT
    PLAID_DIRECT_ITEM_IMPORT_RETURNED_INVALID_MFA = "PLAID_DIRECT_ITEM_IMPORT_RETURNED_INVALID_MFA",

    // RATE_LIMIT_EXCEEDED
    ACCOUNTS_LIMIT = "ACCOUNTS_LIMIT",
    ADDITION_LIMIT = "ADDITION_LIMIT",
    AUTH_LIMIT = "AUTH_LIMIT",
    BALANCE_LIMIT = "BALANCE_LIMIT",
    IDENTITY_LIMIT = "IDENTITY_LIMIT",
    ITEM_GET_LIMIT = "ITEM_GET_LIMIT",
    RATE_LIMIT = "RATE_LIMIT",
    TRANSACTIONS_LIMIT = "TRANSACTIONS_LIMIT",

    // RECAPTCHA_ERROR
    RECAPTCHA_REQUIRED = "RECAPTCHA_REQUIRED",
    RECAPTCHA_BAD = "RECAPTCHA_BAD",

    // OAUTH_ERROR
    INCORRECT_OAUTH_NONCE = "INCORRECT_OAUTH_NONCE",
    OAUTH_STATE_ID_ALREADY_PROCESSED = "OAUTH_STATE_ID_ALREADY_PROCESSED",
}
// ...
export interface LinkEvent {
    eventName: LinkEventName;
    metadata: LinkEventMetadata;
}

export interface LinkEventMetadata {
    accountNumberMask?: string;
    linkSessionId: string;
    mfaType?: string;
    requestId?: string;
    viewName: LinkEventViewName;
    errorCode?: string;
    errorMessage?: string;
    errorType?: string;
    exitStatus?: string;
    institutionId?: string;
    institutionName?: string;
    institutionSearchQuery?: string;
    isUpdateMode?: string;
    matchReason?: string;
    issueId?: string;
    issueDescription?: string;
    issueDetectedAt?: string;
  // see possible values for selection at https://plaid.com/docs/link/web/#link-web-onevent-selection
    selection?: null | string;
    timestamp: string;
    metadata_json: string;
}

export enum LinkEventName {
    AUTO_SELECT_SAVED_INSTITUTION = 'AUTO_SELECT_SAVED_INSTITUTION',
    AUTO_SUBMIT_PHONE = 'AUTO_SUBMIT_PHONE',
    BANK_INCOME_INSIGHTS_COMPLETED = 'BANK_INCOME_INSIGHTS_COMPLETED',
    CLOSE_OAUTH = 'CLOSE_OAUTH',
    CONNECT_NEW_INSTITUTION = 'CONNECT_NEW_INSTITUTION',
    ERROR = 'ERROR',
    EXIT = 'EXIT',
    FAIL_OAUTH = 'FAIL_OAUTH',
    HANDOFF = 'HANDOFF',
    IDENTITY_MATCH_PASSED = 'IDENTITY_MATCH_PASSED',
    IDENTITY_MATCH_FAILED = 'IDENTITY_MATCH_FAILED',
    IDENTITY_VERIFICATION_CLOSE_UI = 'IDENTITY_VERIFICATION_CLOSE_UI',
    IDENTITY_VERIFICATION_CREATE_SESSION = 'IDENTITY_VERIFICATION_CREATE_SESSION',
    IDENTITY_VERIFICATION_FAIL_SESSION = 'IDENTITY_VERIFICATION_FAIL_SESSION',
    IDENTITY_VERIFICATION_FAIL_STEP = 'IDENTITY_VERIFICATION_FAIL_STEP',
    IDENTITY_VERIFICATION_OPEN_UI = 'IDENTITY_VERIFICATION_OPEN_UI',
    IDENTITY_VERIFICATION_PASS_SESSION = 'IDENTITY_VERIFICATION_PASS_SESSION',
    IDENTITY_VERIFICATION_PASS_STEP = 'IDENTITY_VERIFICATION_PASS_STEP',
    IDENTITY_VERIFICATION_PENDING_REVIEW_SESSION = 'IDENTITY_VERIFICATION_PENDING_REVIEW_SESSION',
    IDENTITY_VERIFICATION_PENDING_REVIEW_STEP = 'IDENTITY_VERIFICATION_PENDING_REVIEW_STEP',
    IDENTITY_VERIFICATION_RESUME_SESSION = 'IDENTITY_VERIFICATION_RESUME_SESSION',
    IDENTITY_VERIFICATION_RESUME_UI = 'IDENTITY_VERIFICATION_RESUME_UI',
    IDENTITY_VERIFICATION_START_STEP = 'IDENTITY_VERIFICATION_START_STEP',
    ISSUE_FOLLOWED = 'ISSUE_FOLLOWED',
    LAYER_AUTOFILL_NOT_AVAILABLE = "LAYER_AUTOFILL_NOT_AVAILABLE",
    LAYER_NOT_AVAILABLE = 'LAYER_NOT_AVAILABLE',
    LAYER_READY = 'LAYER_READY',
    MATCHED_CONSENT = 'MATCHED_CONSENT',
    MATCHED_SELECT_INSTITUTION = 'MATCHED_SELECT_INSTITUTION',
    MATCHED_SELECT_VERIFY_METHOD = 'MATCHED_SELECT_VERIFY_METHOD',
    OPEN = 'OPEN',
    OPEN_MY_PLAID = 'OPEN_MY_PLAID',
    OPEN_OAUTH = 'OPEN_OAUTH',
    PLAID_CHECK_PANE = 'PLAID_CHECK_PANE',
    PROFILE_ELIGIBILITY_CHECK_ERROR = 'PROFILE_ELIGIBILITY_CHECK_ERROR',
    PROFILE_ELIGIBILITY_CHECK_READY = 'PROFILE_ELIGIBILITY_CHECK_READY',
    REMEMBER_ME_DISABLED = 'REMEMBER_ME_DISABLED',
    REMEMBER_ME_ENABLED = 'REMEMBER_ME_ENABLED',
    REMEMBER_ME_HOLDOUT = 'REMEMBER_ME_HOLDOUT',
    SEARCH_INSTITUTION = 'SEARCH_INSTITUTION',
    SELECT_ACCOUNT = 'SELECT_ACCOUNT',
    SELECT_AUTH_TYPE = 'SELECT_AUTH_TYPE',
    SELECT_BRAND = 'SELECT_BRAND',
    SELECT_DEGRADED_INSTITUTION = 'SELECT_DEGRADED_INSTITUTION',
    SELECT_DOWN_INSTITUTION = 'SELECT_DOWN_INSTITUTION',
    SELECT_FILTERED_INSTITUTION = 'SELECT_FILTERED_INSTITUTION',
    SELECT_INSTITUTION = 'SELECT_INSTITUTION',
    SELECT_SAVED_ACCOUNT = 'SELECT_SAVED_ACCOUNT',
    SELECT_SAVED_INSTITUTION = 'SELECT_SAVED_INSTITUTION',
    SKIP_SUBMIT_EMAIL = 'SKIP_SUBMIT_EMAIL',
    SKIP_SUBMIT_PHONE = 'SKIP_SUBMIT_PHONE',
    SUBMIT_ACCOUNT_NUMBER = 'SUBMIT_ACCOUNT_NUMBER',
    SUBMIT_CREDENTIALS = 'SUBMIT_CREDENTIALS',
    SUBMIT_DOCUMENTS = 'SUBMIT_DOCUMENTS',
    SUBMIT_DOCUMENTS_ERROR = 'SUBMIT_DOCUMENTS_ERROR',
    SUBMIT_DOCUMENTS_SUCCESS = 'SUBMIT_DOCUMENTS_SUCCESS',
    SUBMIT_EMAIL = 'SUBMIT_EMAIL',
    SUBMIT_MFA = 'SUBMIT_MFA',
    SUBMIT_PHONE = 'SUBMIT_PHONE',
    SUBMIT_OTP = 'SUBMIT_OTP',
    SUBMIT_ROUTING_NUMBER = 'SUBMIT_ROUTING_NUMBER',
    TRANSITION_VIEW = 'TRANSITION_VIEW',
    VERIFY_PHONE = 'VERIFY_PHONE',
    VIEW_DATA_TYPES = 'VIEW_DATA_TYPES',
}

export enum LinkEventViewName {
    ACCEPT_TOS = 'ACCEPT_TOS',
    CONNECTED = 'CONNECTED',
    CONSENT = 'CONSENT',
    CREDENTIAL = 'CREDENTIAL',
    DATA_TRANSPARENCY = 'DATA_TRANSPARENCY',
    DATA_TRANSPARENCY_CONSENT = 'DATA_TRANSPARENCY_CONSENT',
    DOCUMENTARY_VERIFICATION = 'DOCUMENTARY_VERIFICATION',
    ERROR = 'ERROR',
    EXIT = 'EXIT',
    KYC_CHECK = 'KYC_CHECK',
    SELFIE_CHECK = 'SELFIE_CHECK',
    LOADING = 'LOADING',
    MATCHED_CONSENT = 'MATCHED_CONSENT',
    MATCHED_CREDENTIAL = 'MATCHED_CREDENTIAL',
    MATCHED_MFA = 'MATCHED_MFA',
    MFA = 'MFA',
    NUMBERS = 'NUMBERS',
    NUMBERS_SELECT_INSTITUTION = 'NUMBERS_SELECT_INSTITUTION',
    OAUTH = 'OAUTH',
    RECAPTCHA = 'RECAPTCHA',
    RISK_CHECK = 'RISK_CHECK',
    SCREENING = 'SCREENING',
    SELECT_ACCOUNT = 'SELECT_ACCOUNT',
    SELECT_AUTH_TYPE = 'SELECT_AUTH_TYPE',
    SUBMIT_EMAIL = "SUBMIT_EMAIL",
    SUBMIT_PHONE = 'SUBMIT_PHONE',
    VERIFY_PHONE = 'VERIFY_PHONE',
    SELECT_SAVED_INSTITUTION = 'SELECT_SAVED_INSTITUTION',
    SELECT_SAVED_ACCOUNT = 'SELECT_SAVED_ACCOUNT',
    SELECT_BRAND = 'SELECT_BRAND',
    SELECT_INSTITUTION = 'SELECT_INSTITUTION',
    SUBMIT_DOCUMENTS = 'SUBMIT_DOCUMENTS',
    SUBMIT_DOCUMENTS_SUCCESS = 'SUBMIT_DOCUMENTS_SUCCESS',
    SUBMIT_DOCUMENTS_ERROR = 'SUBMIT_DOCUMENTS_ERROR',
    UPLOAD_DOCUMENTS = 'UPLOAD_DOCUMENTS',
    VERIFY_EMAIL = "VERIFY_EMAIL",
    VERIFY_SMS = 'VERIFY_SMS',
}

/// Methods to present Link on iOS.
/// FULL_SCREEN is the converts to UIModalPresentationOverFullScreen on the native side.
/// MODAL will use the default presentation style for iOS which is UIModalPresentationAutomatic.
export enum LinkIOSPresentationStyle {
    FULL_SCREEN = 'FULL_SCREEN',
    MODAL = 'MODAL'
}

export type LinkSuccessListener = (LinkSuccess: LinkSuccess) => void

export type LinkExitListener = (LinkExit: LinkExit) => void

export type LinkOnEventListener = (LinkEvent: LinkEvent) => void